Half Doom World
Afflicted removed as a keyword. Inflict Agony, Karma unaffected.

Doom decays at half if no Vuln, half if no Weak.

Black Bindings changed to "Apply 6(9) Doom if the target has Weak. Apply 2 Weak." Could be plenty of other mechanical triggers here, this seemed to be close-ish to the intentions of the previous text but simpler and without using Afflicted.

Cursed Wail changed to not have a conditional - removing Afflicted bloated the text too much. And it's already akin to a pseudo-Crippling Cloud, as an exhaust card it probably never really needed a conditional. Values currently at 6/6.

Maleficence+ changed to just "twice" for simpler text with the removal of Afflicted. Could do other things here.

Spirit Leech doesn't make sense in the half doom world: Reworked to 7 damage, "gain 1 Reserve next turn". Flying Knee but reserve and a point lower on damage. Could be just about anything we want it to be, this seemed safe enough.